Implement inserting array elements with the `splice()` method (#318)
